Added five new applications.
Added nomacs, PhotoQt, SMPlayer, qimgv, and Vorta.

Also edited the description for Muon, since it's a GUI front-end for APT, so it's only available on Debian and Ubuntu derivatives.

Quick note about a couple of the apps:

nomacs was unmaintained for quite a while and only recently began activity again, so I don't really know when they'll push the next release. Feel free not to add it if you feel like it's outdated or and/or unstable.

qimgv is a special case. The developer lives in Ukraine, and development has been put on hold indefinitely for obvious reasons. However, the dev branch is still active (as of yesterday, at least), but I don't know when or if those changes will be pushed to main any time soon. Totally understandable if you'd prefer not to add this to the list right now.